pyproject.toml
[project]
name = "socialmind"
version = "0.1.0"
description = "Autonomous social media management platform powered by AI"
authors = [{ name = "SocialMind Team" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
	"dspy-ai>=2.0,<3.0",
	"litellm>=1.0,<2.0",
	"openai>=1.0,<2.0",
	"anthropic>=0.25,<0.26",
	"ollama>=0.2,<0.3",
	"httpx>=0.27,<0.28",
	"curl-cffi>=0.7,<0.8",
	"fastapi>=0.111,<0.112",
	"uvicorn[standard]>=0.29,<0.30",
	"pydantic>=2.7,<3.0",
	"pydantic-settings>=2.2,<3.0",
	"python-multipart>=0.0.9,<0.0.10",
	"python-jose[cryptography]>=3.3,<4.0",
	"passlib[bcrypt]>=1.7,<2.0",
	"cryptography>=42.0,<43.0",
	"sqlalchemy[asyncio]>=2.0,<3.0",
	"alembic>=1.13,<2.0",
	"asyncpg>=0.29,<0.30",
	"redis[hiredis]>=5.0,<6.0",
	"celery[redis]>=5.4,<6.0",
	"flower>=2.0,<3.0",
	"mcp>=1.0,<2.0",
	"sse-starlette>=2.0,<3.0",
	"Pillow>=10.0,<11.0",
	"loguru>=0.7,<0.8",
	"typer>=0.12,<0.13",
	"rich>=13.0,<14.0",
	"python-dotenv>=1.0,<2.0",
	"pyotp>=2.9,<3.0",
	"beautifulsoup4>=4.12,<5.0",
	"trafilatura>=1.9,<2.0",
]

[project.optional-dependencies]
dev = [
	"ruff>=0.4,<0.5",
	"mypy>=1.10,<2.0",
	"pytest>=8.0,<9.0",
	"pytest-asyncio>=0.23,<0.24",
	"pytest-playwright>=0.5,<0.6",
	"factory-boy>=3.3,<4.0",
	"pre-commit>=3.7,<4.0",
]
browser = [
	"playwright>=1.44,<2.0",
	"playwright-stealth>=1.0,<2.0",
	"fake-useragent>=1.5,<2.0",
	"python-socks>=2.0,<3.0",
	"aiohttp-socks>=0.8,<0.9",
]
media = [
	"ffmpeg-python>=0.2,<0.3",
	"moviepy>=1.0,<2.0",
]
social-instagram = [
	"instagrapi>=2.0,<3.0",
]
social-reddit = [
	"praw>=7.0,<8.0",
	"asyncpraw>=7.0,<8.0",
]
social-twitter = [
	"tweepy>=4.0,<5.0",
]
social-youtube = [
	"yt-dlp",
	"google-api-python-client>=2.0,<3.0",
	"google-auth-oauthlib>=1.0,<2.0",
	"google-auth-httplib2>=0.2,<0.3",
]

[project.scripts]
socialmind = "socialmind.cli:app"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["socialmind*"]
